requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.15.2
gunicorn==23.0.0
pillow==11.0.0
packaging==24.2
whitenoise==6.8.2